% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{check_rest_version}
\alias{check_rest_version}
\title{Check version of JPred REST interface.}
\usage{

  check_rest_version(host = "http://www.compbio.dundee.ac.uk/jpred4/cgi-bin/rest",
  suffix = "version", silent = FALSE)
}
\arguments{
\item{host}{JPred host address.}

\item{suffix}{Host address suffix.}

\item{silent}{Should the work be done silently? (default = FALSE)}
}
\value{
Version of JPred REST API.
}
\description{
Check version of JPred REST API.
}
\examples{
\dontrun{
## Not run --
check_rest_version()
}
}
