% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{create_jpred_query}
\alias{create_jpred_query}
\title{Create JPRED parameters}
\usage{
create_jpred_query(mode, user_format, file = NULL, seq = NULL,
  skipPDB = TRUE, email = NULL, name = NULL, silent = FALSE)
}
\arguments{
\item{mode}{what mode is being used. See Details.}

\item{user_format}{what format is the data in}

\item{file}{a file to submit}

\item{seq}{alternatively, a sequence in character string}

\item{skipPDB}{should the PDB query be skipped? (default = TRUE)}

\item{email}{for a batch job submission, where to send the results?}

\item{name}{a name for the job.}
}
\value{
list of jpred parameters
}
\description{
Creates a list of JPRED parameters for job submission.
}
