% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpredapir.R
\docType{package}
\name{jpredapir}
\alias{jpredapir}
\alias{jpredapir-package}
\title{\code{jpredapir}: A package for interfacing with JPred server.}
\description{
The \code{jpredapir} package provides functions to submit, check status, and
retrieve results from JPred: A Secondary Structure Prediction Server
(\href{http://www.compbio.dundee.ac.uk/jpred/}{JPred}).
}
\details{
This package contains a collection of functions to interface with JPred server.

For a complete list, use \code{library(help = "jpredapir")}.
}
\section{The \code{jpredapir} package functions}{


\itemize{
\item \code{jpredapir::submit()} Submit job to JPred REST interface.
\item \code{jpredapir::status()} Check job status on JPred REST interface.
\item \code{jpredapir::get_results()} Download results from JPred server.
\item \code{jpredapir::quota()} Check JPred REST interface quota.
\item \code{jpredapir::check_rest_version()} Check version of JPred REST interface.
}
}

\references{
Browse documentation at: \url{https://moseleybioinformaticslab.github.io/jpredapir}

Browse source code at: \url{https://github.com/MoseleyBioinformaticsLab/jpredapir}

Report a bugs at: \url{https://github.com/MoseleyBioinformaticsLab/jpredapir/issues}

JPred Server at: \url{http://www.compbio.dundee.ac.uk/jpred}
}
