% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{quota}
\alias{quota}
\title{Check quota.}
\usage{
quota(email, host = "http://www.compbio.dundee.ac.uk/jpred4/cgi-bin/rest",
  suffix = "quota", silent = FALSE)
}
\arguments{
\item{email}{E-mail address.}

\item{host}{JPred host address.}

\item{suffix}{Host address suffix.}

\item{silent}{Should the work be done silently? (default = FALSE)}
}
\value{
Response.
}
\description{
Check how many jobs you have already submitted on a given day
(out of 1000 maximum allowed jobs per user per day).
}
\examples{

quota(email = "name@domain.com")
}
