% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{resolve_rest_format}
\alias{resolve_rest_format}
\title{Determine JPred REST interface submission format.}
\usage{
resolve_rest_format(mode, user_format)
}
\arguments{
\item{mode}{Submission mode, possible values: "single", "batch", "msa".}

\item{user_format}{Submission format, possible values: "raw", "fasta", "msf", "blc".}
}
\value{
Format for JPred REST interface.
}
\description{
Resolve format of submission to JPred REST interface based on provided mode and user format.
}
