% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{create_jpred_query}
\alias{create_jpred_query}
\title{Create JPred REST interface parameters.}
\usage{
create_jpred_query(rest_format, file = NULL, seq = NULL, skipPDB = TRUE,
  email = NULL, name = NULL, silent = FALSE)
}
\arguments{
\item{rest_format}{Format for JPred REST interface.}

\item{file}{File path to a file with the job input (sequence or msa).}

\item{seq}{Alternatively, amino acid sequence passed as string of single-letter code without spaces, e.g. --seq=ATWFGTHY}

\item{skipPDB}{Should the PDB query be skipped? (default = TRUE)}

\item{email}{For a batch job submission, where to send the results?}

\item{name}{A name for the job.}

\item{silent}{Should the work be done silently? (default = FALSE)}
}
\value{
Formatted string with JPred parameters.
}
\description{
Create a string with JPred parameters for job submission.
}
