% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{status}
\alias{status}
\title{Check status.}
\usage{
status(job_id, results_dir_path = NULL, extract = FALSE,
  max_attempts = 10, wait_interval = 60, silent = FALSE,
  host = "http://www.compbio.dundee.ac.uk/jpred4/cgi-bin/rest",
  jpred4 = "http://www.compbio.dundee.ac.uk/jpred4")
}
\arguments{
\item{job_id}{Job id.}

\item{results_dir_path}{Directory path where to save results if job is finished.}

\item{extract}{Extract or not results into directory (default = FALSE).}

\item{max_attempts}{Maximum number of attempts to check job status (default = 10).}

\item{wait_interval}{Wait interval before retrying to check job status in seconds (default = 60).}

\item{silent}{Should the work be done silently? (default = FALSE).}

\item{host}{JPred host address.}

\item{jpred4}{JPred address for results retrieval.}
}
\value{
Response.
}
\description{
Check status of the submitted job.
}
