% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{submit}
\alias{submit}
\title{Submit job to JPred interface.}
\usage{
submit(mode, user_format, file = NULL, seq = NULL, skipPDB = TRUE,
  email = NULL, name = NULL, silent = FALSE,
  host = "http://www.compbio.dundee.ac.uk/jpred4/cgi-bin/rest",
  suffix = "job")
}
\arguments{
\item{mode}{Submission mode, possible values: "single", "batch", "msa".}

\item{user_format}{Submission format, possible values: "raw", "fasta", "msf", "blc".}

\item{file}{File path to a file with the job input (sequence or msa).}

\item{seq}{Alternatively, amino acid sequence passed as string of single-letter code without spaces, e.g. --seq=ATWFGTHY}

\item{skipPDB}{Should the PDB query be skipped? (default = TRUE)}

\item{email}{For a batch job submission, where to send the results?}

\item{name}{A name for the job.}

\item{silent}{Should the work be done silently? (default = FALSE)}

\item{host}{JPred host address.}

\item{suffix}{Host address suffix.}
}
\value{
Response.
}
\description{
Submit job to JPred REST API.
}
\examples{
\dontrun{
## Not run ---
example_path <- system.file("exampledata", package = "jpredapir")
submit(mode = "single", user_format = "raw", seq = "MQVWPIEGIKKFETLSYLPP")
submit(mode = "single", user_format = "raw", 
       file = file.path(example_path, "single_raw.example"))
submit(mode = "single", user_format = "fasta", 
       file = file.path(example_path, "single_fasta.example"))
submit(mode = "batch", user_format = "fasta", 
       file = file.path(example_path, "batch_fasta.example"), email = "name@domain.com")
submit(mode = "msa", user_format = "fasta", 
       file = file.path(example_path, "msa_fasta.example"), email = "name@domain.com")
submit(mode = "msa", user_format = "msf", 
       file = file.path(example_path, "msa_msf.example"), email = "name@domain.com")
submit(mode = "msa", user_format = "blc", 
       file = file.path(example_path, "msa_blc.example"), email = "name@domain.com")
}
}
